+++ app/src/main/AndroidManifest.xml
@@ -32,29 +32,6 @@
             android:name=".MainActivity"
             android:exported="true" />
 
-        <activity
-            android:name=".HomeActivity"
-            android:exported="true" />
-
-        <activity
-            android:name=".ChatActivity"
-            android:exported="true" />
-
-        <activity
-            android:name=".SwipeActivity"
-            android:exported="true" />
-
-        <activity
-            android:name=".ProfileActivity"
-            android:exported="true" />
-
-        <activity
-            android:name=".StudentActivity"
-            android:exported="true" />
-
-        <activity
-            android:name=".RegularActivity"
-            android:exported="true" />
 
     </application>
