+++ app/src/main/java/com/example/myapplication/RegularFragment.kt
@@ -1,37 +1,47 @@
 package com.example.myapplication
 
 import android.content.ContentValues.TAG
-import android.content.Intent
 import android.os.Bundle
+import android.text.method.ScrollingMovementMethod
 import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import android.widget.*
-import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatEditText
+import androidx.fragment.app.Fragment
 import com.google.firebase.FirebaseApp
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
 
-class RegularActivity : AppCompatActivity() {
+class RegularFragment : Fragment() {
 
     private lateinit var db: FirebaseFirestore
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_regular)
-        FirebaseApp.initializeApp(this)
+        FirebaseApp.initializeApp(requireContext())
         db = FirebaseFirestore.getInstance()
-        loadData()
-        val saveButton = findViewById<Button>(R.id.save_button)
-        saveButton.setOnClickListener {
-            saveProfile()
-            Toast.makeText(this, "Account Information Saved.", Toast.LENGTH_LONG).show()
-            val home = Intent(this, HomeActivity::class.java)
-            startActivity(home)
-        }
     }
 
-    private fun loadData() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_regular, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val saveButton = view.findViewById<Button>(R.id.save_button)
+        saveButton.setOnClickListener {
+            saveProfile()
+            Toast.makeText(context, "Account Information Saved.", Toast.LENGTH_LONG).show()
+        }
+
         val auth = FirebaseAuth.getInstance()
         val currentUser = auth.currentUser
         if (currentUser != null) {
@@ -40,7 +50,7 @@ class RegularActivity : AppCompatActivity() {
                 .addOnSuccessListener { documentSnapshot ->
                     if (documentSnapshot.exists()) {
                         val profile = documentSnapshot.data
-                        val fullNameTextView = findViewById<TextView>(R.id.fullName)
+                        val fullNameTextView = view.findViewById<TextView>(R.id.fullName)
                         val firstName = profile?.get("firstName").toString()
                         val lastName = profile?.get("lastName").toString()
                         fullNameTextView.text = "$firstName $lastName"
@@ -61,41 +71,42 @@ class RegularActivity : AppCompatActivity() {
                     if (documentSnapshot.exists()) {
                         val accountInfo = documentSnapshot.data
 
-                        val addressSpinner = findViewById<Spinner>(R.id.address)
+                        val addressSpinner = view.findViewById<Spinner>(R.id.address)
                         val addressAdapter = addressSpinner.adapter as ArrayAdapter<String>
                         val addressPosition =
                             addressAdapter.getPosition(accountInfo?.get("hasAddress").toString())
                         addressSpinner.setSelection(addressPosition)
 
-                        val jobTextView = findViewById<TextView>(R.id.job)
+                        val jobTextView = view.findViewById<TextView>(R.id.job)
                         jobTextView.text = accountInfo?.get("job").toString()
 
-                        val rentView = findViewById<TextView>(R.id.rent)
+                        val rentView = view.findViewById<TextView>(R.id.rent)
                         rentView.text = accountInfo?.get("rent").toString()
 
-                        val crimeSpinner = findViewById<Spinner>(R.id.hasCriminalHistory)
+                        val crimeSpinner = view.findViewById<Spinner>(R.id.hasCriminalHistory)
                         val crimeAdapter = crimeSpinner.adapter as ArrayAdapter<String>
                         val crimePosition = crimeAdapter.getPosition(
                             accountInfo?.get("hasCriminalHistory").toString()
                         )
                         crimeSpinner.setSelection(crimePosition)
 
-                        val crimeTextView = findViewById<TextView>(R.id.criminalHistory)
+                        val crimeTextView = view.findViewById<TextView>(R.id.criminalHistory)
                         crimeTextView.text = accountInfo?.get("criminalHistory").toString()
+                        crimeTextView.movementMethod = ScrollingMovementMethod()
 
-                        val petsSpinner = findViewById<Spinner>(R.id.pets)
+                        val petsSpinner = view.findViewById<Spinner>(R.id.pets)
                         val petsAdapter = petsSpinner.adapter as ArrayAdapter<String>
                         val petsPosition =
                             petsAdapter.getPosition(accountInfo?.get("pets").toString())
                         petsSpinner.setSelection(petsPosition)
 
-                        val kidsSpinner = findViewById<Spinner>(R.id.kids)
+                        val kidsSpinner = view.findViewById<Spinner>(R.id.kids)
                         val kidsAdapter = kidsSpinner.adapter as ArrayAdapter<String>
                         val kidsPosition =
                             kidsAdapter.getPosition(accountInfo?.get("kids").toString())
                         kidsSpinner.setSelection(kidsPosition)
 
-                        val smokerSpinner = findViewById<Spinner>(R.id.smoker)
+                        val smokerSpinner = view.findViewById<Spinner>(R.id.smoker)
                         val smokerAdapter = smokerSpinner.adapter as ArrayAdapter<String>
                         val smokerPosition =
                             smokerAdapter.getPosition(accountInfo?.get("smoker").toString())
@@ -112,21 +123,23 @@ class RegularActivity : AppCompatActivity() {
     }
 
     private fun saveProfile() {
-        val addressTypeSpinner = findViewById<Spinner>(R.id.address)
+        val view = requireView()
+
+        val addressTypeSpinner = view.findViewById<Spinner>(R.id.address)
         val address = addressTypeSpinner.selectedItem.toString()
-        val jobField = findViewById<AppCompatEditText>(R.id.job)
+        val jobField = view.findViewById<AppCompatEditText>(R.id.job)
         val job = jobField.text.toString()
-        val rentField = findViewById<AppCompatEditText>(R.id.rent)
+        val rentField = view.findViewById<AppCompatEditText>(R.id.rent)
         val rent = rentField.text.toString()
-        val crimeSpinner = findViewById<Spinner>(R.id.hasCriminalHistory)
+        val crimeSpinner = view.findViewById<Spinner>(R.id.hasCriminalHistory)
         val hasCriminalHistory = crimeSpinner.selectedItem.toString()
-        val crimeField = findViewById<AppCompatEditText>(R.id.criminalHistory)
+        val crimeField = view.findViewById<AppCompatEditText>(R.id.criminalHistory)
         val criminalHistory = crimeField.text.toString()
-        val petsTypeSpinner = findViewById<Spinner>(R.id.pets)
+        val petsTypeSpinner = view.findViewById<Spinner>(R.id.pets)
         val pets = petsTypeSpinner.selectedItem.toString()
-        val kidsTypeSpinner = findViewById<Spinner>(R.id.kids)
+        val kidsTypeSpinner = view.findViewById<Spinner>(R.id.kids)
         val kids = kidsTypeSpinner.selectedItem.toString()
-        val smokerTypeSpinner = findViewById<Spinner>(R.id.smoker)
+        val smokerTypeSpinner = view.findViewById<Spinner>(R.id.smoker)
         val smoker = smokerTypeSpinner.selectedItem.toString()
 
 
@@ -175,6 +188,5 @@ class RegularActivity : AppCompatActivity() {
             Log.w(TAG, "Error: current user is null")
         }
     }
-
 
 }
