+++ app/src/main/java/com/example/myapplication/SignUpActivity.kt
@@ -8,8 +8,6 @@ import androidx.appcompat.app.AppCompatActivity
 import com.example.myapplication.databinding.ActivityMainBinding
 import com.example.myapplication.databinding.ActivitySignUpBinding
 import com.google.firebase.auth.FirebaseAuth
-import com.example.myapplication.R
-
 
 class SignUpActivity: AppCompatActivity() {
 
@@ -24,6 +22,7 @@ class SignUpActivity: AppCompatActivity() {
 
         firebaseAuth = FirebaseAuth.getInstance()
 
+
         binding.textView.setOnClickListener {
             val intent = Intent(this, SignInActivity::class.java)
             startActivity(intent)
@@ -34,6 +33,7 @@ class SignUpActivity: AppCompatActivity() {
             val pass = binding.passET.text.toString()
             val confirmPass = binding.confirmPassEt.text.toString()
 
+
             if(email.isNotEmpty() && pass.isNotEmpty() && confirmPass.isNotEmpty()){
                 if (pass == confirmPass) {
 
@@ -41,8 +41,6 @@ class SignUpActivity: AppCompatActivity() {
                         if (it.isSuccessful){
                             val intent = Intent(this, SignInActivity::class.java)
                             startActivity(intent)
-                            finish() // finish the SignUpActivity so the user cannot navigate back to it
-
                         }else{
                             Toast.makeText(this, it.exception.toString(), Toast.LENGTH_SHORT).show()
                         }
@@ -61,4 +59,3 @@ class SignUpActivity: AppCompatActivity() {
     }
 
 }
-
