+++ app/src/main/res/layout/fragment_regular.xml
