+++ app/src/main/res/layout/fragment_swipe.xml
@@ -5,13 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/BottomNav"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        app:menu="@menu/bottom_menu"/>
-
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/HomeToolbar"
         android:layout_width="match_parent"
